requirements.txt
selenium==4.15.0
beautifulsoup4
requests
python-dotenv